Fix thread safety of the DriverConfigTemplate class
A single DriverConfigTemplate class is created for a capability template file. When running multi-threaded, we could run into a case where a template is reevaluated by multiple threads, and they all access the same instance of the EnhancedCapabilities class. When running a series of CapabilityOverrider functions, we risk one thread corrupting the capabilities of another thread. Adding a synchronized block around all DriverConfigTemplate functions will ensure that each thread gets its own copy of the EnhancedCapabilities.